chore(#501): removed old reduction-process
Removed old reduction-process-code, which doesn't
work anymore anyway. I wanted to test another way,
but basically failed, because the performance was
much worse and the result not better. Left the new
code disabled in the back-propagation. Maybe it will
be tested again in the future, when bigger test-cases
are available.