fix: fix invalid dataset-download-request
When downloading a dataset with a too high request
amount of rows, the download failed with the wrong
error-message. This was fixed by an additonal check,
which cuts the number if too big and downloads only
the available values, even when the requested amount
of rows was much bigger. A too big row-offset is now
checked by an additional if-condition to give a
correct error-message in this case.