feat(#454): support multiple outputs
Result-datasets were changed. When making a request-task
against a cluster with multiple outputs, each output-hexagon
generates its own dataset. Because of this, the uuid of the
request-task is no longer the uuid of the resulting dataset,
because there can be multiple datasets as result. So the
datasets now have a 'task_uuid'-field, which contains the
uuid of the request-task.