revert(#533): remove cluster-mode again
The cluster-mode, which was an original idea of the old
backend c++ implmentation, was removed again. While
implementing the train and request endpoint for direct
interation, i found out, that the new workflow of the
cluster provides more oportunities. So in the new version
the direct-input can be mixed with the task-input without
any conflicts.

Signed-off-by: Tobias Anker <tobias.anker@kitsunemimi.moe>