feat(#561): allow multiple epochs
The creation of train-tasks was updated to allow the
definition of a number of epochs for the task to run
multiple iterations over the same dataset. Because this
was already in mind while refactoring code from c++ into
rust, this feature was already prepeared in many places,
so the changes here were quite minimal.

Signed-off-by: Tobias Anker <tobias.anker@kitsunemimi.moe>