fix: fix duplicate project and user id
For projects and users the id in the database was the
primary key. So it was only able to exist one time each.
Problem was, that the entries are not deleted, but only
marked as deleted. So a project or user id, which one
time existed in the database, couldn't be added again, even
it was marked as deleted. So the ids are no longer
primary keys, but instead it is manually checked, that the
id is only one time at the same time marked as active.

Signed-off-by: Tobias Anker <tobias.anker@kitsunemimi.moe>