perf(#577): restructure task-processing
Changed the task-processing to be more like in the old
c++ version, there the worker-thread, which process the
last output, also triggers the next cycle. This improves
the performance in small cycles quite a lot, because there
is not idle between the end of a cycle and the start of
a new one.

Signed-off-by: Tobias Anker <tobias.anker@kitsunemimi.moe>