fix(#647): fixed deadlock in bigger cluster
In bigger clusters with multiple core-hexagons,
there were deadlocks between the core-blocks of these
possible, in case that the backpropagation started
before all paths reached the output. The problem
was solves by some reordering of locks and replacing
the lock in backpropagation, for transfering axons
backward, by a try_lock with retry-loop.

Signed-off-by: Tobias Anker <tobias.anker@kitsunemimi.moe>