feat(#651): allow cpu-limitation
Added config to allow to limit the number of
cpu-threads to use for the worker-threads.
The new config is per default 0, which means
there is no limitation and all cpu-threads, expect
the 2 reserved ones, are allowed to be used as
worker-threads.
Also removed the old memory-limitation-config,
because there is no pre-allocation of memory anymore.

Signed-off-by: Tobias Anker <tobias.anker@kitsunemimi.moe>