fix(#592): fix randomly failing uploads
When uploading datasets to the server, they are cached as
file first, before the file is read and converted. Sometimes
reading this cache-file failed, especially in the CI-pipeline.
The reason seems to be a missing file-sync after writing the
file, so it was not completely on the disc before it tried
to read the file again. Added sync to fix this problem.
Also updated error-log in this case a bit.

Signed-off-by: Tobias Anker <tobias.anker@kitsunemimi.moe>