fix(#592): fix randomly failing sdk-api-test
The last test int eh sdk-api-test could randomly failed, if
the result doesn't match completely the expected value. The
used example for the last test seems not the optimal one and
can in rare cases be not correctly trained. This should be
solved now by adding more train-iterations for this case
to ensure, it is correclty trained.

Signed-off-by: Tobias Anker <tobias.anker@kitsunemimi.moe>