fix: fix commit-linter
While creating the commitlint-file to add the ignore-rule
for the dependency-bot commits, I didn't know, that a ruleset
has to be added as well, even this was not necessary before
the file was created. So all commits were ignored from the check.
This was fixed by adding the conventional-ruleset now again.

Signed-off-by: Tobias Anker <tobias.anker@kitsunemimi.moe>