fix(#676): fix accuracy problem in bigger tests
In slightly bigger tests with some weather data
the output was quite broken. After investigation
it appeared, that the handling of the overflow of
potential to the next section was not calculated
correctly, which resulted in the cut in the potential
in case one section is not enough. The calculation
was fixed, which resulted in quite good results
in the first tests so far.

Signed-off-by: Tobias Anker <tobias.anker@kitsunemimi.moe>