refactor(#731): refactored big parts of the api
The REST-api endpoints of ryokan and sakura got a big
refactoring to reduce redundant code and make the
api-function better readable. Some of them became simply
too long and complex, which made the changes in context
of the new onsen-component much more difficult.
Fixed checkpoints and dataset-check for the new
onsen-backend.
Also tested to add some rustdoc comments to some
of the new functions.

Signed-off-by: Tobias Anker <tobias.anker@kitsunemimi.moe>