feat(#731): use wireguard-connection in kubenetes-setup
In order to get an encrypted connection from sakura and
ryokan to the onsen, in the kubernetes-setup a wireguard
connection is used now. The configs for this have to be
generated with a new python-script. Only supports one
onsen-node in the kubernetes-setup at the moment, because
the wireguard-configs are at the moment hard-coded for
only one onsen.
Updated install-docu for this and removed the old longhorn
installation.

Signed-off-by: Tobias Anker <tobias.anker@kitsunemimi.moe>