feat(#431): made all datasets encrypted
All datasets uploaded to the ryokan will be
encrypted with an autogenerated omamori-secret,
before the dataset is uploaded to the onsen.
Also resulting datasets of request-tasks are now
be encrypted before they are uploaded.
I also got a bit carried away to add already an
new learning for cleaner error-handling to the code,
even this should have been done in a separate commit.

Signed-off-by: Tobias Anker <tobias.anker@kitsunemimi.moe>