feat(#720): random host-selection
Until now, always the first sakura- and onsen-host
were choosen from a list in case of scheduling. This
means that effectively only one host could be used.
This was now updated that a random host from the list
is choosen. Not the best scheduling, but at least now
are more than one sakura- and onsen-host possible.

Signed-off-by: Tobias Anker <tobias.anker@kitsunemimi.moe>